Use leading whitespace in Toit code snippets.
We sometimes divide code into several triple-backtick
sections, separated by commentary.  In this case all the
lines in later sections can be indented, and this should
be preserved for cut-and-paste.

This change also alters the syntax colorer for Toit so
that it assumes we are inside a function like main if
the first line is indented.  This gives much nicer colouring
for many examples on docs.toit.io.